backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["av3n93rz <daniel.raffer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
ruff = "^0.2.2"
uvicorn = "^0.27.1"
langchain = "^0.1.9"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.7.4"
pgvector = "^0.2.5"
psycopg2-binary = "^2.9.9"
tiktoken = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
